Crate gmsol_sdk

Re-exports

pub use error::Error;

pub use gmsol_utils as core;

pub use gmsol_programs as programs;

pub use gmsol_solana_utils as solana_utils;

Modules

builders

Instruction Builders.

constants

Constants.

error

Error type.

market

Market calculations.

market_graph

Maintains a graph structured with MarketModel as edges.

model

Model support.

pda

Functions for constructing Program Derived Addresses.

position

Position calculations.

utils

Utils.

Structs

AtomicGroup

A group of instructions that are expected to be executed in the same transaction.

ParallelGroup

A group of atomic instructions that can be executed in parallel.

TransactionGroup

Transaction Group.

Enums

SolanaUtilsError

Error type.

Traits

IntoAtomicGroup

A trait representing types that can be converted into AtomicGroups.

Type Aliases

Result

Result type.