gmsol_sdk::model::position

Trait PositionExt

pub trait PositionExt<const DECIMALS: u8>: Position<DECIMALS> {
    // Provided methods
    fn will_collateral_be_sufficient(
        &self,
        prices: &Prices<Self::Num>,
        delta: &CollateralDelta<Self::Num>,
    ) -> Result<WillCollateralBeSufficient<Self::Signed>, Error> { ... }
    fn collateral_price<'a>(
        &self,
        prices: &'a Prices<Self::Num>,
    ) -> &'a Price<Self::Num> { ... }
    fn collateral_value(
        &self,
        prices: &Prices<Self::Num>,
    ) -> Result<Self::Num, Error> { ... }
    fn pnl_value(
        &self,
        prices: &Prices<Self::Num>,
        size_delta_usd: &Self::Num,
    ) -> Result<(Self::Signed, Self::Signed, Self::Num), Error> { ... }
    fn validate(
        &self,
        prices: &Prices<Self::Num>,
        should_validate_min_position_size: bool,
        should_validate_min_collateral_usd: bool,
    ) -> Result<(), Error> { ... }
    fn check_liquidatable(
        &self,
        prices: &Prices<Self::Num>,
        should_validate_min_collateral_usd: bool,
    ) -> Result<Option<LiquidatableReason>, Error> { ... }
    fn position_price_impact(
        &self,
        size_delta_usd: &Self::Signed,
    ) -> Result<Self::Signed, Error> { ... }
    fn capped_positive_position_price_impact(
        &self,
        index_token_price: &Price<Self::Num>,
        size_delta_usd: &Self::Signed,
    ) -> Result<Self::Signed, Error> { ... }
    fn capped_position_price_impact(
        &self,
        index_token_price: &Price<Self::Num>,
        size_delta_usd: &Self::Signed,
    ) -> Result<(Self::Signed, Self::Num), Error> { ... }
    fn pending_borrowing_fee_value(&self) -> Result<Self::Num, Error> { ... }
    fn pending_funding_fees(&self) -> Result<FundingFees<Self::Num>, Error> { ... }
    fn position_fees(
        &self,
        collateral_token_price: &Price<Self::Num>,
        size_delta_usd: &Self::Num,
        is_positive_impact: bool,
        is_liquidation: bool,
    ) -> Result<PositionFees<Self::Num>, Error> { ... }
}

Extension trait for Position with utils.

Provided Methods

fn will_collateral_be_sufficient( &self, prices: &Prices<Self::Num>, delta: &CollateralDelta<Self::Num>, ) -> Result<WillCollateralBeSufficient<Self::Signed>, Error>

Check that whether the collateral will be sufficient after paying the given realized_pnl and applying delta_size.

Returns the remaining collateral value if sufficient, None otherwise.

Returns Err if failed to finish the calculation.

fn collateral_price<'a>( &self, prices: &'a Prices<Self::Num>, ) -> &'a Price<Self::Num>

Get collateral price.

fn collateral_value( &self, prices: &Prices<Self::Num>, ) -> Result<Self::Num, Error>

Get collateral value.

fn pnl_value( &self, prices: &Prices<Self::Num>, size_delta_usd: &Self::Num, ) -> Result<(Self::Signed, Self::Signed, Self::Num), Error>

Calculate the pnl value when decreased by the given delta size.

Returns (pnl_value, uncapped_pnl_value, size_delta_in_tokens)

fn validate( &self, prices: &Prices<Self::Num>, should_validate_min_position_size: bool, should_validate_min_collateral_usd: bool, ) -> Result<(), Error>

Validate the position.

fn check_liquidatable( &self, prices: &Prices<Self::Num>, should_validate_min_collateral_usd: bool, ) -> Result<Option<LiquidatableReason>, Error>

Check if the position is liquidatable.

Return LiquidatableReason if it is liquidatable, None otherwise.

fn position_price_impact( &self, size_delta_usd: &Self::Signed, ) -> Result<Self::Signed, Error>

Get position price impact.

fn capped_positive_position_price_impact( &self, index_token_price: &Price<Self::Num>, size_delta_usd: &Self::Signed, ) -> Result<Self::Signed, Error>

Get position price impact usd and cap the value if it is positive.

fn capped_position_price_impact( &self, index_token_price: &Price<Self::Num>, size_delta_usd: &Self::Signed, ) -> Result<(Self::Signed, Self::Num), Error>

Get capped position price impact usd.

Compare to PositionExt::capped_positive_position_price_impact, this method will also cap the negative impact and return the difference before capping.

fn pending_borrowing_fee_value(&self) -> Result<Self::Num, Error>

Get pending borrowing fee value of this position.

fn pending_funding_fees(&self) -> Result<FundingFees<Self::Num>, Error>

Get pending funding fees.

fn position_fees( &self, collateral_token_price: &Price<Self::Num>, size_delta_usd: &Self::Num, is_positive_impact: bool, is_liquidation: bool, ) -> Result<PositionFees<Self::Num>, Error>

Calculates the PositionFees generated by changing the position size by the specified size_delta_usd.