gmsol_sdk::model::position

Trait PositionStateMut

pub trait PositionStateMut<const DECIMALS: u8>: PositionState<DECIMALS> {
    // Required methods
    fn collateral_amount_mut(&mut self) -> &mut Self::Num;
    fn size_in_usd_mut(&mut self) -> &mut Self::Num;
    fn size_in_tokens_mut(&mut self) -> &mut Self::Num;
    fn borrowing_factor_mut(&mut self) -> &mut Self::Num;
    fn funding_fee_amount_per_size_mut(&mut self) -> &mut Self::Num;
    fn claimable_funding_fee_amount_per_size_mut(
        &mut self,
        is_long_collateral: bool,
    ) -> &mut Self::Num;
}

Mutable access to the position state.

Required Methods

fn collateral_amount_mut(&mut self) -> &mut Self::Num

Get a mutable reference to the collateral amount.

fn size_in_usd_mut(&mut self) -> &mut Self::Num

Get a mutable reference to the size (in USD) of the position.

fn size_in_tokens_mut(&mut self) -> &mut Self::Num

Get a mutable reference to the size (in tokens) of the position.

fn borrowing_factor_mut(&mut self) -> &mut Self::Num

Get a mutable reference to last borrowing factor applied by the position.

fn funding_fee_amount_per_size_mut(&mut self) -> &mut Self::Num

Get a mutable reference to the funding fee amount per size.

fn claimable_funding_fee_amount_per_size_mut( &mut self, is_long_collateral: bool, ) -> &mut Self::Num

Get a mutable reference to claimable funding fee amount per size of the given collateral.