gmsol_sdk::model

Trait BalanceExt

pub trait BalanceExt: Balance {
    // Provided methods
    fn long_usd_value(&self, price: &Self::Num) -> Result<Self::Num, Error> { ... }
    fn short_usd_value(&self, price: &Self::Num) -> Result<Self::Num, Error> { ... }
    fn pool_delta_with_amounts(
        &self,
        long_token_delta_amount: &Self::Signed,
        short_token_delta_amount: &Self::Signed,
        long_token_price: &Self::Num,
        short_token_price: &Self::Num,
    ) -> Result<PoolDelta<Self::Num>, Error> { ... }
    fn pool_delta_with_values(
        &self,
        delta_long_token_usd_value: Self::Signed,
        delta_short_token_usd_value: Self::Signed,
        long_token_price: &Self::Num,
        short_token_price: &Self::Num,
    ) -> Result<PoolDelta<Self::Num>, Error> { ... }
    fn merge<B>(&self, short: B) -> Merged<&Self, B>
       where B: Balance { ... }
    fn amount(&self, is_long: bool) -> Result<Self::Num, Error> { ... }
}

Extension trait for Balance with utils.

Provided Methods

fn long_usd_value(&self, price: &Self::Num) -> Result<Self::Num, Error>

Get the long amount value in USD.

fn short_usd_value(&self, price: &Self::Num) -> Result<Self::Num, Error>

Get the short amount value in USD.

fn pool_delta_with_amounts( &self, long_token_delta_amount: &Self::Signed, short_token_delta_amount: &Self::Signed, long_token_price: &Self::Num, short_token_price: &Self::Num, ) -> Result<PoolDelta<Self::Num>, Error>

Get pool value information after applying delta.

fn pool_delta_with_values( &self, delta_long_token_usd_value: Self::Signed, delta_short_token_usd_value: Self::Signed, long_token_price: &Self::Num, short_token_price: &Self::Num, ) -> Result<PoolDelta<Self::Num>, Error>

Get pool value information after applying delta.

fn merge<B>(&self, short: B) -> Merged<&Self, B>
where B: Balance,

Merge the amounts with other Balance.

The result Balance will consider the total amounts (long_amount + short_amount) of self as the long amount, and the total amount of short as the short amount.

fn amount(&self, is_long: bool) -> Result<Self::Num, Error>

Get amount by side.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.