gmsol_sdk::model

Trait BaseMarketExt

pub trait BaseMarketExt<const DECIMALS: u8>: BaseMarket<DECIMALS> {

    // Provided methods
    fn pool_value_without_pnl_for_one_side(
        &self,
        prices: &Prices<Self::Num>,
        is_long: bool,
        maximize: bool,
    ) -> Result<Self::Num, Error> { ... }
    fn open_interest(&self) -> Result<Merged<&Self::Pool, &Self::Pool>, Error> { ... }
    fn open_interest_in_tokens(
        &self,
    ) -> Result<Merged<&Self::Pool, &Self::Pool>, Error> { ... }
    fn pnl(
        &self,
        index_token_price: &Price<Self::Num>,
        is_long: bool,
        maximize: bool,
    ) -> Result<Self::Signed, Error> { ... }
    fn pnl_factor_with_pool_value(
        &self,
        prices: &Prices<Self::Num>,
        is_long: bool,
        maximize: bool,
    ) -> Result<(Self::Signed, Self::Num), Error> { ... }
    fn pnl_factor(
        &self,
        prices: &Prices<Self::Num>,
        is_long: bool,
        maximize: bool,
    ) -> Result<Self::Signed, Error> { ... }
    fn validate_pool_amount(&self, is_long_token: bool) -> Result<(), Error> { ... }
    fn pnl_factor_exceeded(
        &self,
        prices: &Prices<Self::Num>,
        kind: PnlFactorKind,
        is_long: bool,
    ) -> Result<Option<PnlFactorExceeded<Self::Num>>, Error> { ... }
    fn validate_pnl_factor(
        &self,
        prices: &Prices<Self::Num>,
        kind: PnlFactorKind,
        is_long: bool,
    ) -> Result<(), Error> { ... }
    fn validate_max_pnl(
        &self,
        prices: &Prices<Self::Num>,
        long_kind: PnlFactorKind,
        short_kind: PnlFactorKind,
    ) -> Result<(), Error> { ... }
    fn reserved_value(
        &self,
        index_token_price: &Price<Self::Num>,
        is_long: bool,
    ) -> Result<Self::Num, Error> { ... }
    fn validate_reserve(
        &self,
        prices: &Prices<Self::Num>,
        is_long: bool,
    ) -> Result<(), Error> { ... }
    fn expected_min_token_balance_excluding_collateral_amount_for_one_token_side(
        &self,
        is_long_side: bool,
    ) -> Result<Self::Num, Error> { ... }
    fn total_collateral_amount_for_one_token_side(
        &self,
        is_long_side: bool,
    ) -> Result<Self::Num, Error> { ... }

}

Extension trait for BaseMarket.

Provided Methods

fn pool_value_without_pnl_for_one_side( &self, prices: &Prices<Self::Num>, is_long: bool, maximize: bool, ) -> Result<Self::Num, Error>

Get the usd value of primary pool without pnl for one side.

fn open_interest(&self) -> Result<Merged<&Self::Pool, &Self::Pool>, Error>

Get total open interest as a Balance.

fn open_interest_in_tokens( &self, ) -> Result<Merged<&Self::Pool, &Self::Pool>, Error>

Get total open interest in tokens as a merged Balance.

The long amount is the total long open interest in tokens, while the short amount is the total short open interest in tokens.

fn pnl( &self, index_token_price: &Price<Self::Num>, is_long: bool, maximize: bool, ) -> Result<Self::Signed, Error>

Get total pnl of the market for one side.

fn pnl_factor_with_pool_value( &self, prices: &Prices<Self::Num>, is_long: bool, maximize: bool, ) -> Result<(Self::Signed, Self::Num), Error>

Get pnl factor with pool value.

fn pnl_factor( &self, prices: &Prices<Self::Num>, is_long: bool, maximize: bool, ) -> Result<Self::Signed, Error>

Get pnl factor.

fn validate_pool_amount(&self, is_long_token: bool) -> Result<(), Error>

Validate (primary) pool amount.

fn pnl_factor_exceeded( &self, prices: &Prices<Self::Num>, kind: PnlFactorKind, is_long: bool, ) -> Result<Option<PnlFactorExceeded<Self::Num>>, Error>

Get the excess of pending pnl.

Return Some if the pnl factor is exceeded the given kind of pnl factor.

fn validate_pnl_factor( &self, prices: &Prices<Self::Num>, kind: PnlFactorKind, is_long: bool, ) -> Result<(), Error>

Validate pnl factor.

fn validate_max_pnl( &self, prices: &Prices<Self::Num>, long_kind: PnlFactorKind, short_kind: PnlFactorKind, ) -> Result<(), Error>

Validate max pnl.

fn reserved_value( &self, index_token_price: &Price<Self::Num>, is_long: bool, ) -> Result<Self::Num, Error>

Get reserved value.

fn validate_reserve( &self, prices: &Prices<Self::Num>, is_long: bool, ) -> Result<(), Error>

Validate reserve.

fn expected_min_token_balance_excluding_collateral_amount_for_one_token_side( &self, is_long_side: bool, ) -> Result<Self::Num, Error>

Expected min token balance excluding collateral amount.

Notes

Note that “one token side” here means calculating based on half of the side. For markets where the long token and short token are different, there is no ambiguity. However, if the long token and short token are the same, choosing the long token side will result in a value that is not actually the total amount of the long token, but rather the total amount belonging to the long token side (often only half of it).

For example, if both the long token and the short token are WSOL, and the liquidity pool has a total of 1000 WSOL. Then, in a typical pool implementation, the long token side of the liquidity pool has only 500 WSOL, while the short token side also has 500 WSOL. In this case, this function will only consider one side, taking into account only 500 WSOL in the calculation.

fn total_collateral_amount_for_one_token_side( &self, is_long_side: bool, ) -> Result<Self::Num, Error>

Get total collateral amount for one token side.

Notes

Note that “one token side” here means calculating based on half of the side. (See also expected_min_token_balance_excluding_collateral_amount_for_one_token_side).